+++ src/ui/message-actions/show-large-message.js
@@ -1,14 +1,14 @@
 /**
- * Simple action handler for the `layer-open-expanded-view` action.  Opens a dialog showing the model that the action is performed against
+ * Simple action handler for the `layer-show-large-message` action.  Opens a dialog showing the model that the action is performed against
  *
- * @class Layer.UI.MessageActions.OpenExpandedView
+ * @class Layer.UI.MessageActions.ShowLargeMessage
  */
 
 import { register } from './index';
 import { logger } from '../../utils';
 
-const openExpandedView = ({ messageViewer, model, data }) => {
-  const dialog = document.createElement('layer-message-viewer-expanded');
+const showLargeView = ({ messageViewer, model, data }) => {
+  const dialog = document.createElement('layer-large-message-viewer');
   dialog.model = model;
   dialog.openActionData = data;
   let node = messageViewer;
@@ -25,4 +25,4 @@ const openExpandedView = ({ messageViewer, model, data }) => {
   }
 };
 
-register('layer-open-expanded-view', openExpandedView);
+register('layer-show-large-message', showLargeView);